# front-end/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi - Easy3vs3",
    "description": "OpenApi documentation for Easy3vs3 project",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "samuele",
      "email": "samueletoscano@hotmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local ENV"
    },
    {
      "url": "http://localhost:8080",
      "description": "Prod ENV"
    }
  ],
  "paths": {
...
    "/api/generate/semi-final/{tournament-name}": {
      "post": {
        "tags": ["ROUND "],
        "summary": "Generate Semi Finals",
        "description": "Generate semi finals for a tournament.",
        "operationId": "generateSemiFinals",
        "parameters": [
          {
            "name": "tournament-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Easy3vs3Auth": []
          }
        ]
      }
    },
    "/api/generate/quarter-final/{tournament-name}": {
      "post": {
        "tags": ["ROUND "],
        "summary": "Generate Quarter Finals",
        "description": "Generate quarter finals for a tournament.",
        "operationId": "generateQuarterFinals",
        "parameters": [
          {
            "name": "tournament-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Easy3vs3Auth": []
          }
        ]
      }
    },
    "/api/generate/final/{tournament-name}": {
      "post": {
        "tags": ["ROUND "],
        "summary": "Generate Final Match",
        "description": "Generate final match for a tournament.",
        "operationId": "generateFinal",
        "parameters": [
          {
            "name": "tournament-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Easy3vs3Auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "via": {
            "type": "string"
          },
          "civico": {
            "type": "string"
          },
          "cap": {
            "type": "string"
          },
          "townName": {
            "type": "string"
          },
          "siglaProvince": {
            "type": "string"
          }
        }
      },
      "Competition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "coverUrl": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "referees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Referee"
            }
          },
          "teams": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Team"
            }
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          }
        }
      },
      "Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "homeTeam": {
            "$ref": "#/components/schemas/Team"
          },
          "awayTeam": {
            "$ref": "#/components/schemas/Team"
          },
          "homePoints": {
            "type": "integer",
            "format": "int32"
          },
          "awayPoints": {
            "type": "integer",
            "format": "int32"
          },
          "winner": {
            "$ref": "#/components/schemas/Team"
          },
          "status": {
            "type": "string",
            "enum": ["SCHEDULED", "STARTED", "FINISHED"]
          },
          "round": {
            "type": "string",
            "enum": ["OCTAVEFINAL", "QUARTERFINAL", "SEMIFINAL", "FINAL"]
          },
          "tournament": {
            "$ref": "#/components/schemas/Competition"
          }
        }
      },
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "courtName": {
            "type": "string"
          }
        }
      },
      "Player": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "point": {
            "type": "integer",
            "format": "int32"
          },
          "nickname": {
            "type": "string"
          },
          "sigla": {
            "type": "string"
          },
          "gamesPlayed": {
            "type": "integer",
            "format": "int32"
          },
          "roleInTheGame": {
            "type": "string",
            "enum": ["PLAYER", "REFEREE"]
          },
          "teamName": {
            "type": "string"
          }
        }
      },
      "Referee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "nickname": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["PLAYER", "REFEREE"]
          }
        }
      },
      "Team": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "players": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Player"
            }
          },
          "logo": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "captain": {
            "$ref": "#/components/schemas/Player"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Easy3vs3Auth": {
        "type": "http",
        "description": "Here, if you are authorized, you can have the access to all endopints",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}