# front-end/src/app/api/models/player.ts
/* tslint:disable */
/* eslint-disable */
export interface Player {
  age?: number;
  dateOfBirth: string;
  gamesPlayed?: number;
  id: string;
  name: string;
  nickname: string;
  point?: number;
  roleInTheGame?: 'PLAYER' | 'REFEREE';
  sigla: string;
  surname: string;
  teamName: string;
}